requirements.txt
pandas>=1.3.0
numpy>=1.19.0
openai>=1.0.0
pyvis>=0.2.0
networkx>=2.6.0
sentence-transformers>=2.2.0
openpyxl>=3.0.0
ipython>=7.0.0
matplotlib>=3.3.0
scikit-learn>=0.24.0
biopython>=1.80